# directive tag injector: rename applicationId to version
